Fix grid and reduce subchunkopt

+++ src/engine/engine.h
@@ -12,8 +12,8 @@ extern bool DEBUG_ON;
 
 extern byte gameboard[VSCREEN_HEIGHT][VSCREEN_WIDTH];
 
-/* 32-bit subchunk_t means 32x32 subchunks filling VSCREEN */
-typedef uint32_t subchunk_t;
+/* 16-bit subchunk_t means 16x16 subchunks filling VSCREEN */
+typedef uint16_t subchunk_t;
 
 #define SUBCHUNK_SIZE	(8 * sizeof(subchunk_t))
 #define SUBCHUNK_HEIGHT (VSCREEN_HEIGHT / SUBCHUNK_SIZE)
